moondream-worker/requirements.txt
runpod>=1.7.6         
transformers>=4.40.0
accelerate>=0.23.0
pillow>=10.0.0
einops
pyvips-binary==8.16.0
pyvips==2.2.3
